Avoid signed bitfield

+++ lib/libriscv/cpu_dispatch.cpp
@@ -332,7 +332,7 @@ INSTRUCTION(RV32I_BC_JAL, rv32i_jal) {
 	if constexpr (false) {
 		fprintf(stderr, "JAL PC 0x%lX => 0x%lX\n", long(pc), long(pc+fi.offset));
 	}
-	NEXT_BLOCK(fi.offset, true);
+	NEXT_BLOCK((int32_t)fi.offset, true);
 }
 
 /** UNLIKELY INSTRUCTIONS **/
